+++ pyramid_oereb/contrib/data_sources/standard/sources/plr.py
@@ -9,7 +9,6 @@ from sqlalchemy import text, or_
 
 from pyramid_oereb import Config
 from pyramid_oereb.core import b64
-from pyramid_oereb.core.records.availability import AvailabilityRecord
 from pyramid_oereb.core.records.image import ImageRecord
 from pyramid_oereb.core.records.plr import EmptyPlrRecord
 from pyramid_oereb.core.sources import BaseDatabaseSource
@@ -101,32 +100,16 @@ class DatabaseSource(BaseDatabaseSource, PlrBaseSource):
                 as key and text as value.
         """
         config_parser = StandardThemeConfigParser(**kwargs)
-        models = config_parser.get_models()
+        self.models = config_parser.get_models()
         bds_kwargs = {
-            'model': models.Geometry,
+            'model': self.models.Geometry,
             'db_connection': kwargs.get('source').get('params').get('db_connection')
         }
 
         BaseDatabaseSource.__init__(self, **bds_kwargs)
         PlrBaseSource.__init__(self, **kwargs)
 
-        self.legend_entry_model = models.LegendEntry
-        availability_model = models.Availability
-
-        self.availabilities = []
-
-        session = self.get_session()
-
-        try:
-
-            availabilities_from_db = session.query(availability_model).all()
-            for availability in availabilities_from_db:
-                self.availabilities.append(
-                    AvailabilityRecord(availability.fosnr, available=availability.available)
-                )
-
-        finally:
-            session.close()
+        self.legend_entry_model = self.models.LegendEntry
 
     def from_db_to_legend_entry_record(self, legend_entry_from_db):
         theme = Config.get_theme_by_code_sub_code(legend_entry_from_db.theme)
@@ -370,9 +353,15 @@ class DatabaseSource(BaseDatabaseSource, PlrBaseSource):
     def get_document_records(self, params, public_law_restriction_from_db):
         documents_from_db = []
         if not hasattr(public_law_restriction_from_db, 'legal_provisions'):
-            raise AttributeError('The public_law_restriction implementation of type {} has no '
-                                 'legal_provisions attribute. Check the model implementation.'
-                                 .format(type(public_law_restriction_from_db)))
+            raise AttributeError(
+                'The public_law_restriction implementation of type {} has no '
+                'legal_provisions attribute. Check the model implementation. Theme: {}, ID: {}, {}' .format(
+                    type(public_law_restriction_from_db),
+                    self._plr_info.get('code'),
+                    public_law_restriction_from_db.__weakref__,
+                    dir(public_law_restriction_from_db)
+                )
+            )
         for legal_provision in public_law_restriction_from_db.legal_provisions:
             documents_from_db.append(legal_provision.document)
         document_records = self.from_db_to_document_records(documents_from_db)
@@ -503,8 +492,9 @@ class DatabaseSource(BaseDatabaseSource, PlrBaseSource):
             bbox (shapely.geometry.base.BaseGeometry): The bbox to search the records.
         """
         # Check if the plr is marked as available
-        if self._is_available(real_estate):
+        if Config.availability_by_theme_code_municipality_fosnr(self._plr_info['code'], real_estate.fosnr):
             session = self.get_session()
+
             try:
                 if session.query(self._model_).count() == 0:
                     # We can stop here already because there are no items in the database
@@ -545,19 +535,3 @@ class DatabaseSource(BaseDatabaseSource, PlrBaseSource):
                 Config.get_theme_by_code_sub_code(self._plr_info['code']),
                 has_data=False
             )]
-
-    def _is_available(self, real_estate):
-        """
-        Checks if the topic is available for the specified real estate.
-
-        Args:
-            real_estate (pyramid_oereb.lib.records.real_estate.RealEstateRecord): The real
-                estate in its record representation.
-
-        Returns:
-             bool: True if the topic is available, false otherwise.
-        """
-        for availability in self.availabilities:
-            if int(real_estate.fosnr) == int(availability.fosnr) and not availability.available:
-                return False
-        return True